# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Burrow JSON API
  description: |
    The hand-written canonical OpenAPI v3 description of the Burrow JSON API
    (unified across v0.2.0, v0.3.0, and v0.4.0).

    Coverage is locked by `TestOpenAPI_RouteCoverage` in
    `internal/api/openapi_test.go`: every route the chi mux answers must
    appear in this document or the build fails. Health probes (`/healthz`,
    `/readyz`) and the doc-serving endpoints (`/api/v1/openapi.{yaml,json}`)
    are excluded by convention.

    Schemas are intentionally minimal — the goal is shape coverage for SDK
    code-generators (TypeScript-fetch, Go), not exhaustive validation. Full
    schemas can be filled in incrementally without invalidating downstream
    SDKs (additive only).
  version: 0.4.0
  license:
    name: MIT
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8080
        description: Burrow control-plane host.
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: burrow_session
      description: |
        Session cookie issued by POST /api/v1/auth/login. State-changing
        cookie-authed requests MUST also send the burrow_csrf double-submit
        header (`X-CSRF-Token`).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: bua_<token>
      description: |
        Automation bearer token (`Authorization: Bearer bua_...`). Bypasses
        CSRF (the token IS the CSRF defense). Permissions are the
        intersection of the user's current role and the set declared at
        mint time.
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit token for cookie-authed state-changing requests.
        Value mirrors the burrow_csrf cookie.
  schemas:
...
    AIModelTarget:
      type: object
      description: One provider and model a synthetic model can be sent to.
      required: [provider, model]
      properties:
        dialect:
          type: string
          enum: [openai, anthropic]
          description: >
            The API format the target is used for. It must equal the
            provider's api_format; left out of a request, it is filled in
            from the provider.
        provider:
          type: string
          description: Slug of an existing provider.
        model:
          type: string
          minLength: 1
          maxLength: 200
          description: The provider's own model id. No control characters.
    AIModel:
      type: object
      description: >
        A synthetic model: a name that clients send as "model" on the
        gateway's endpoints and that resolves to an ordered list of provider
        targets per dialect. A request reaches only targets of its own
        dialect.
      properties:
        name:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]{1,62}$'
          description: >
            Never contains "/". "v1" is reserved, and a name cannot equal a
            provider slug.
        description: { type: string, maxLength: 500 }
        enabled: { type: boolean }
        fallback_on_rate_limit: { type: boolean }
        attempt_timeout_s: { type: integer, minimum: 1, maximum: 600, default: 60 }
        total_timeout_s:
          type: integer
          minimum: 1
          maximum: 600
          default: 120
          description: Not shorter than attempt_timeout_s.
        targets:
          type: array
          minItems: 1
          maxItems: 16
          description: >
            At most 8 per dialect. Within one dialect the order is the order
            the targets are tried in.
          items: { $ref: '#/components/schemas/AIModelTarget' }
        dialects:
          type: array
          readOnly: true
          description: The formats the model is served in, sorted; derived from the targets.
          items: { type: string, enum: [anthropic, openai] }
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
...
    AIProviderModel:
      type: object
      description: One entry of a provider's stored model list.
      properties:
        id: { type: string, maxLength: 200, example: google/gemini-x }
        display_name: { type: string, maxLength: 200 }
        context_length: { type: integer, format: int64, minimum: 0 }
        synced_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
...
  responses:
    Unauthorized:
      description: Missing / invalid session or bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but the caller lacks permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - CookieAuth: []
    CSRFToken: []
  - BearerAuth: []
paths:
...
  /api/v1/ai/providers/{slug}/models:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [models]
      summary: List a provider's stored models
      description: |
        The list GET /ai/<provider>/v1/models answers from. Returns 404 when
        no provider has that slug or its backing service is not visible to
        the caller.
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Models ordered by id.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIProviderModel'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [models]
      summary: Add one model id by hand (admin)
      description: Works for either kind of provider. Adding an id that is already listed changes nothing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: No control characters, no surrounding space.
            example: { id: glm-5.1 }
      responses:
        '204': { description: Model added. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [models]
      summary: Remove one model id (admin)
      parameters:
        - name: id
          in: query
          required: true
          description: The model id. A query parameter because model ids contain "/".
          schema: { type: string, minLength: 1, maxLength: 200 }
      responses:
        '204': { description: Model removed. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/ai/models:
    get:
      tags: [models]
      summary: List synthetic models
      description: Every synthetic model, ordered by name. Any signed-in caller may read them.
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: The models.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AIModel' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [models]
      summary: Create a synthetic model (admin or ai:configure:any)
      description: |
        The body is at most 16 KiB and may carry only the fields of the
        schema; dialects and the timestamps are not accepted. enabled left
        out means true; a timeout left out or 0 means its default. Every
        target must name an existing provider, and its dialect, when given,
        must equal that provider's api_format.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AIModel' }
            example:
              name: burrow-intelligence
              targets:
                - { dialect: openai, provider: zai, model: glm-5.1 }
                - { dialect: anthropic, provider: zai-anthropic, model: glm-5.1 }
      responses:
        '201':
          description: Model created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AIModel' }
        '400':
          description: >
            The body is not valid; the message says why (name, an unknown
            field, an unknown provider, a provider of another format, the
            number of targets, a timeout).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Another model has the name.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/ai/models/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [models]
      summary: Get one synthetic model
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: The model.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AIModel' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [models]
      summary: Replace a synthetic model (admin or ai:configure:any)
      description: |
        The body replaces the model; targets is the new list. Two fields
        keep their stored value when left out: name and enabled. A different
        name renames the model. Gateway keys that name the old name in
        allowed_models are not rewritten and no longer reach the model.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AIModel' }
      responses:
        '200':
          description: Model updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AIModel' }
        '400':
          description: The body is not valid; the message says why.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Another model has the new name. Nothing was changed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [models]
      summary: Delete a synthetic model (admin or ai:configure:any)
      description: |
        Requests for the name are refused from then on. Gateway keys keep
        the name in allowed_models; it matches nothing until a model of that
        name exists again.
      responses:
        '204': { description: Model deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...